http: add to_reply() in httpd::redirect_exception
This function allows for the creation of an `http::reply`
from a thrown `redirect_exception`, with a url for the `Location`
header and an optional `Retry-After` timeout value.

Reverted some changes from `v24.1.x` that didn't make it to the upstream
`seastar` in order to move changes to `v24.2.x`.